Renamed m_competition to baseline.
