+++ src/gluonts/model/baseline/_naive_2_predictor.py
