+++ test/model/baseline/test_r_code_compliance_of_naive_2.py
@@ -19,7 +19,7 @@ from pathlib import Path
 
 # First-party imports
 from gluonts.dataset.common import Dataset
-from gluonts.model.m_competitions import naive_2
+from gluonts.model.baseline import naive_2
 
 
 # DISCLAIMER:
